skill-matching-fastapi/requirements.txt
FastAPI
uvicorn
sentence-transformers
psycopg2-binary
pgvector
python-dotenv